wardrobe wizards/requirements.txt
flask
flask-cors
python-dotenv
pillow
opencv-python
numpy
groq
